requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.9.2
httpx==0.27.0
pypdf==4.3.1
python-dotenv==1.0.1

